requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.0.0
    # via
    #   darbiadev-ups
    #   httpcore
certifi==2023.7.22
    # via
    #   darbiadev-ups
    #   httpcore
    #   httpx
click==8.1.7
    # via typer
darbiadev-shipping[ups]==0.5.1
    # via -r requirements/requirements.in
darbiadev-shipping-types==0.2.0
    # via -r requirements/requirements.in
darbiadev-ups==0.7.0
    # via darbiadev-shipping
et-xmlfile==1.1.0
    # via openpyxl
h11==0.14.0
    # via
    #   darbiadev-ups
    #   httpcore
httpcore==0.18.0
    # via
    #   darbiadev-ups
    #   httpx
httpx==0.25.0
    # via darbiadev-ups
idna==3.4
    # via
    #   anyio
    #   darbiadev-ups
    #   httpx
numpy==1.26.2
    # via pandas
openpyxl==3.1.2
    # via -r requirements/requirements.in
pandas==2.1.4
    # via -r requirements/requirements.in
pydantic==2.5.2
    # via pydantic-settings
pydantic-core==2.14.5
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via pydantic-settings
pytz==2023.3.post1
    # via pandas
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   darbiadev-ups
    #   httpcore
    #   httpx
typer==0.9.0
    # via -r requirements/requirements.in
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2023.3
    # via pandas
